// src/core/entities/users/authentication.enum.ts
export enum PermissionsValues {
  GUEST_MODE = "GUEST_MODE",

  READ_FEATURE_FLAG = "READ_FEATURE_FLAG",
  CREATE_FEATURE_FLAG = "CREATE_FEATURE_FLAG",
  UPDATE_FEATURE_FLAG = "UPDATE_FEATURE_FLAG",
  REMOVE_FEATURE_FLAG = "INTERNAL_ADMIN_REMOVE_FEATURE_FLAG",

  READ_CONFIGURATION = "READ_CONFIGURATION",
  CREATE_CONFIGURATION = "CREATE_CONFIGURATION",
  UPDATE_CONFIGURATION = "UPDATE_CONFIGURATION",
  REMOVE_CONFIGURATION = "INTERNAL_ADMIN_REMOVE_CONFIGURATION",

  READ_SUBSCRIPTION_PLAN = "READ_SUBSCRIPTION_PLAN",
  CREATE_SUBSCRIPTION_PLAN = "ADMIN_CREATE_SUBSCRIPTION_PLAN",
  UPDATE_SUBSCRIPTION_PLAN = "ADMIN_UPDATE_SUBSCRIPTION_PLAN",
  REMOVE_SUBSCRIPTION_PLAN = "INTERNAL_ADMIN_REMOVE_SUBSCRIPTION_PLAN",
}